ai-dev-kit/databricks-tools-core/requirements.txt
databricks-sdk>=0.67.0
pydantic>=2.0.0
